refactor: remove unused codes
